requirements.txt
Flask
Flask-Cors
scikit-learn
requests
python-dotenv
pandas # 모델 훈련(train_emotion_model.py) 시 필요